refactor: extract component from EntryContent (#617)
* refactor: extract EntryTitle component from EntryContent

* refactor: extract SupportCreator component from EntryContent
